fixed some errors

+++ programming_paradigm/robust_division_calculator.py
@@ -1,6 +1,6 @@
 def safe_divide(numerator, denominator):
-    numerator = int(numerator)
-    denominator = int(denominator)
+    numerator = float(numerator)
+    denominator = float(denominator)
     try:
         result = numerator / denominator
     except ZeroDivisionError:
